new nav added, audit log updated, make payment now creates partial and auto due invoice, weekly exam now records float value as marks, time stamp is now fixed to bd time

+++ database/migrations/2025_11_14_064630_create_weekly_exam_marks_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->enum('class_level', ['hsc_1', 'hsc_2']);
             $table->enum('section', ['science', 'humanities', 'business_studies']);
             $table->date('exam_date');
-            $table->integer('marks_obtained');
+            $table->decimal('marks_obtained', 8, 2);
             $table->integer('max_marks')->default(100);
             $table->foreignId('recorded_by')->constrained('users')->cascadeOnDelete();
             $table->text('remarks')->nullable();
